Python/requirements.txt
Appium-Python-Client==0.20
attrs==15.2.0
BeautifulSoup
cffi==1.5.2
cryptography==1.3.1
cssselect==0.9.1
Django==1.8.5
ecdsa==0.13
enum34==1.1.2
Fabric==1.11.1
idna==2.1
ipaddress==1.0.16
lxml==3.6.0
paramiko==1.17.0
pillow==3.1.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
pymongo==3.2.2
pyOpenSSL==16.0.0
pytesseract==0.1.6
queuelib==1.4.2
redis==2.10.3
reportlab==3.3.0
requests
Scrapy==1.0.5
scrapy-crawlera==1.1.0
scrapy-mongodb==0.9.1
scrapy-splash==0.6
scrapyjs==0.2
selenium==2.49.0
service-identity==16.0.0
six==1.10.0
Twisted==16.0.0
virtualenv==15.0.1
w3lib==1.13.0
zc.buildout==2.5.1
zope.interface==4.1.3
